Added joining and removing events synchronization
TODO: editing still doesn't propogate changes to joined
Has anyone noticed a weird bug in gcal today?
When I make for a event today, in the month view it shows sunday
but in the agenda view, week view and the event view (when you click on it, the date is correct).